withdrew

+++ src/controller/brand/index.js
@@ -61,28 +61,30 @@ const getBrand = async (req, res) => {
             error: "key is not valid"
         })
     }
-
 }
 const deleteBrand = async (req, res) => {
     const { key } = req.headers
     const { name } = req.body
     if (key == process.env.key_Brand) {
         await brandSchema.findOne({ name: name }).then(async (brand) => {
-            await categorySchema.findOne({ name: brand.category }).then(async (category) => {
-                await categorySchema.findOneAndUpdate({ name: category.name }, { totalBrand: category.totalBrand - 1 }).then(async () => {
-                    await brandSchema.findOneAndDelete({ name: name }).then((data) => {
-                        res.status(200).send({
-                            data
+            if (brand == null) {
+                res.status(200).send({
+                    message: "This brand is not registered"
+                })
+            }
+            else {
+                await categorySchema.findOne({ name: brand.category }).then(async (category) => {
+                    await categorySchema.findOneAndUpdate({ name: category.name }, { totalBrand: category.totalBrand - 1 }).then(async () => {
+                        await brandSchema.findOneAndDelete({ name: name }).then((data) => {
+                            res.status(200).send({
+                                data
+                            })
                         })
                     })
                 })
-            })
+            }
+
         })
-
-
-
-
-
     }
     else {
         res.status(201).send({
@@ -105,10 +107,23 @@ const updateBrand = async (req, res) => {
             error: "key is not valid"
         })
     }
+}
 
+const getBrandData = async (req, res) => {
+    const { key, name } = req.headers
+    if (key == process.env.key_Brand) {
+        await brandSchema.find({ name: { $in: name } }).then((data) => {
+            res.status(200).send({
+                data
+            })
+        })
+    }
+    else {
+        res.status(201).send({
+            error: "key is not valid"
+        })
+    }
 }
 
 
-
-
-module.exports = { createBrand, getAllBrand, getBrand, updateBrand, deleteBrand }
+module.exports = { getBrandData, createBrand, getAllBrand, getBrand, updateBrand, deleteBrand }
